adding functionality to restrict update and delete to only the owner

+++ app/routers/post.py
@@ -45,6 +45,9 @@ def delete_post(id : int,db : Session = Depends(database.get_db), current_user =
     if not post_query.first():
         raise HTTPException(status_code=status.HTTP_404_NOT_FOUND,
                             detail=f'post with id : {id} does not exist')
+    if post_query.first().owner_id != current_user.id:
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN,
+                            detail=f'Not authorized to perform requested action')
     post_query.delete(synchronize_session=False)
     db.commit()
     return Response(status_code=status.HTTP_204_NO_CONTENT)
@@ -58,6 +61,9 @@ def update_post(id : int, post:schemas.PostCreate,db : Session = Depends(databas
     if not post_to_update_query.first():
         raise HTTPException(status_code=status.HTTP_404_NOT_FOUND,
                             detail=f'post with id : {id} does not exist')
+    if post_to_update_query.first().owner_id != current_user.id:
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN,
+                            detail=f'Not authorized to perform requested action')
     post_to_update_query.update(post.model_dump(),synchronize_session=False)
     db.commit()
     return post_to_update_query.first()
